% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosg.R
\name{cosg}
\alias{cosg}
\title{Accurate and fast cell marker gene identification with COSG}
\usage{
cosg(
  object,
  groups = "all",
  assay = "RNA",
  slot = "data",
  mu = 1,
  n_genes_user = 100
)
}
\arguments{
\item{assay}{Assay to use in marker gene identification}

\item{slot}{Slot to pull data from}

\item{mu}{The penalty factor to penalize gene expression in cells not belonging to the cluster of interest}
  
\item{remove_lowly_expressed}{If TRUE, genes that express a percentage of target cells smaller than a specific value (`expressed_pct`) are not considered as marker genes for the target cells. The default value is TRUE.}

\item{expressed_pct}{When `remove_lowly_expressed` is set to TRUE, genes that express a percentage of target cells smaller than a specific value (`expressed_pct`) are not considered as marker genes for the target cells. The default value for `expressed_pct`
        is 0.1 (10%).}
  
\item{n_genes_user}{Number of top ranked genes returned in the result}
}
\value{
A list containing two dataframes for ranked marker genes' names and scores, respectively
}
\description{
Marker gene identification for cell groups in a given dataset.
}
\examples{
suppressMessages(library(Seurat))
data('pbmc_small',package='Seurat')
# Check cell groups:
table(Idents(pbmc_small))
#######
# Run COSG:
marker_cosg <- cosg(
 pbmc_small,
 groups='all',
 assay='RNA',
 slot='data',
 mu=1,
 n_genes_user=100)
#######
# Check the marker genes:
 head(marker_cosg$names)
 head(marker_cosg$scores)
}
