% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_small}
\alias{pbmc_small}
\title{A small example Seurat object for testing}
\format{
A Seurat object with:
\describe{
\item{assays}{RNA assay with counts and data layers}
\item{meta.data}{Cell metadata including cell type annotations}
}
}
\source{
SeuratObject package
}
\usage{
pbmc_small
}
\description{
A subsetted version of a 10X Genomics PBMC dataset, containing 80 cells
and 230 genes. This is the same dataset as SeuratObject::pbmc_small.
}
\examples{
data(pbmc_small)
pbmc_small
}
\keyword{datasets}
